pyproject.toml
[project]
name = "agentic-framework"
version = "0.2.3"
description = "An opinionated framework for building sophisticated AI Agents"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "html2text>=2024.2.26",
    "jinja2>=3.1.5",
    "pyyaml>=6.0.2",
    "swarm",
    "thespian>=4.0.0",
    "litellm>=1.63.14",
    "typer>=0.15.1",
    "cryptography>=44.0.0",
    "ray>=2.42.0",
    "ray[serve]>=2.42.0",
    "aiofiles>=24.1.0",
    "pydantic>=2.10.6",
    "pandas>=2.2.3",
    "fastapi>=0.115.8",
    "setuptools>=75.8.0",
    "requests>=2.32.3",
    "python-magic>=0.4.27",
    "httpx>=0.28.1",
    "sse-starlette>=2.2.1",
    "sqlmodel>=0.0.22",
    "numpy>=2.1.3",
    "textract<1.6.4",
    "pypdf2>=3.0.1",
]

[project.optional-dependencies]
# Dashboard package
dashboard = []

# Streamlit package
streamlit = [
    "streamlit>=1.41.1",
]

# RAG packages
rag = [
    "weaviate-client>=4.10.4",
    "chonkie[semantic]>=0.4.1",
    "transformers>=4.48.3",
    "torch>=2.6.0",
    "grpcio>=1.70.0",
    "fastembed>=0.5.1",
    "extract-msg<=0.29.0",
]

# Tool groups
airbnb = [
    "icalendar>=6.1.3",
]

browser-use = [
    "browser-use>=0.1.40",
    "playwright>=1.51.0",
    "langchain>=0.3.22",
    "langchain-google-genai>=2.1.2"
]

database = [
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.26",
]

duckduckgo = [
    "duckduckgo-search>=7.5.5",
]

github = [
    "gitpython>=3.1.44",
]

google-news = [
    "google-news-feed>=1.1.0",
    "googlenewsdecoder>=0.1.7",
]

image-generator = [
   "boto3>=1.36.19",
]

imap = [
    "beautifulsoup4>=4.13.4",
]

mcp = [
    "mcp[cli]>=1.5.0",
]

meeting-baas = [
    "agentic-framework[rag]",
    "openai>=1.75.0",
    "sqlalchemy>=2.0.26",
]

playwright = [
    "playwright>=1.51.0",
]

text-to-speech = [
    "pydub>=0.25.1",
]

all-tools = [
    "agentic-framework[airbnb,browser-use,database,duckduckgo,github,google-news,image-generator,imap,mcp,meeting-baas,playwright,text-to-speech]",
]

all = [
    "agentic-framework[all-tools,rag,dashboard,streamlit]",
]

dev = [
    "pytest",
    "pytest-asyncio>=0.25.3",
    "debugpy>=1.8.0",
    "black",
    "mkdocs-material>=9.6.4",
    "mkdocs-swagger-ui-tag>=0.6.11",
    "mike>=2.1.3"
]

[project.scripts]
agentic = "agentic.cli:app"

[tool.hatch.version]
path = "agentic/__init__.py"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src", "agentic_examples" = "examples"}
packages = [
    "agentic", 
    "agentic.custom_models",
    "agentic.tools",
    "agentic.tools.utils",
    "agentic.swarm",
    "agentic.streamlit",
    "agentic.dashboard",
    "agentic.db",
    "agentic.utils",
    "agentic_examples"
]
include-package-data = true
